+++ src/TextureWebAppChrome.js
@@ -4,27 +4,12 @@ import TextureAppChrome from './TextureAppChrome'
 import { VfsStorageClient, HttpStorageClient, InMemoryDarBuffer } from './dar'
 
 export default class TextureWebAppChrome extends TextureAppChrome {
-  _loadArchive (archiveId, context, cb) {
-    let storage = this._getStorage(this.props.storageType)
-    let buffer = new InMemoryDarBuffer()
-    let ArchiveClass = this._getArchiveClass()
-    let archive = new ArchiveClass(storage, buffer, context, this._config)
-    // Don't catch exception in debug mode
-    if (this.props.debug) {
-      archive.load(archiveId, cb)
-    } else {
-      try {
-        archive.load(archiveId, cb)
-      } catch (err) {
-        this.setState({
-          error: err
-        })
-        console.error(err)
-      }
-    }
+  _getBuffer () {
+    return new InMemoryDarBuffer()
   }
 
-  _getStorage (storageType) {
+  _getStorage () {
+    let storageType = this.props.storageType
     if (storageType === 'vfs') {
       return new VfsStorageClient(vfs, this._getDefaultDataFolder())
     } else {
@@ -54,8 +39,4 @@ export default class TextureWebAppChrome extends TextureAppChrome {
       event.stopPropagation()
     }
   }
-
-  _getArchiveClass () { throw new Error('This method is abstract') }
-
-  _getDefaultDataFolder () { throw new Error('This method  is abstract') }
 }
